+++ web_ui/src/providers/dataset-import-to-new-project-provider/dataset-import-to-new-project-provider.component.tsx
@@ -24,10 +24,10 @@ import {
     DatasetImportToNewProjectItem,
 } from '../../core/datasets/dataset.interface';
 import { useDatasetImportQueries } from '../../core/datasets/hooks/use-dataset-import-queries.hook';
-import { useLocalStorageDatasetImport } from '../../core/datasets/hooks/use-local-storage-dataset-import.hook';
 import { useFeatureFlags } from '../../core/feature-flags/hooks/use-feature-flags.hook';
 import QUERY_KEYS from '../../core/requests/query-keys';
 import { useApplicationServices } from '../../core/services/application-services-provider.component';
+import { useLocalStorageDatasetImport } from '../../features/dataset-import/hooks/use-local-storage-dataset-import.hook';
 import { NOTIFICATION_TYPE } from '../../notification/notification-toast/notification-type.enum';
 import { useNotification } from '../../notification/notification.component';
 import { getRandomDistinctColor } from '../../pages/create-project/components/distinct-colors';
